Move jwtkey to an exp package
